add validation

+++ app/Http/Controllers/Portal/BlogDataController.php
@@ -7,6 +7,7 @@ use Illuminate\Http\Request;
 use Carbon\Carbon;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Validation\ValidationException;
 
 class BlogDataController extends Controller
 {
@@ -15,19 +16,19 @@ class BlogDataController extends Controller
     }
 
     public function store(Request $request) {
-        // dd($request->all());
+
+        $request->validate([
+            'title'         => 'required|string|max:255',
+            'banner'        => 'nullable|image|mimes:jpg,jpeg,png|max:2048',
+            'description'   => 'nullable|string',
+            'featured'      => 'nullable|string',
+            'publish_by'    => 'nullable|string|max:255',
+            'publish_date'  => 'nullable|date',
+        ]);
+
         DB::beginTransaction();
         try {
-            $request->validate([
-                'title'         => 'required|string|max:255',
-                'banner'        => 'nullable|image|mimes:jpg,jpeg,png|max:2048',
-                'description'   => 'nullable|string',
-                'featured'      => 'nullable|string',
-                'publish_by'    => 'nullable|string|max:255',
-                'publish_date'  => 'nullable|date',
-            ]);
-
-            $validatedData = $request->only(['title', 'description', 'featured', 'publish_by', 'publish_date']);
+            $validatedData = $request->only(['title', 'banner', 'description', 'featured', 'publish_by', 'publish_date']);
 
             if ($request->filled('publish_date')) {
                 $validatedData['publish_date'] = Carbon::parse($request->publish_date)->format('Y-m-d');
@@ -41,8 +42,12 @@ class BlogDataController extends Controller
             }
             
             Blog::create($validatedData);
+
             DB::commit();
             return redirect()->route('blog.list')->with('success', 'Blog submitted successfully!');
+        } catch (ValidationException $e) {
+            DB::rollBack();
+            return redirect()->back()->withErrors($e->errors())->withInput();
         } catch (\Exception $e) {
             DB::rollBack();
             return redirect()->back()->with('error', 'Failed to submit Blog: ' . $e->getMessage());
@@ -75,20 +80,23 @@ class BlogDataController extends Controller
         return view('blog.add-edit', compact('blog'));
     }
 
-    public function update(Request $request, $id) {
+    public function update(Request $request, $id) 
+    {
+
+        $request->validate([
+            'title'        => 'required|string|max:255',
+            'banner'       => 'nullable|image|mimes:jpg,jpeg,png|max:2048',
+            'description'      => 'nullable|string',
+            'featured'      => 'nullable|string',
+            'publish_by'   => 'nullable|string|max:255',
+            'publish_date' => 'nullable|date',
+        ]);
+
         DB::beginTransaction();
         try {
-            $request->validate([
-                'title'        => 'required|string|max:255',
-                'banner'       => 'nullable|image|mimes:jpg,jpeg,png|max:2048',
-                'description'      => 'nullable|string',
-                'featured'      => 'nullable|string',
-                'publish_by'   => 'nullable|string|max:255',
-                'publish_date' => 'nullable|date',
-            ]);
-            
+
             $blog = Blog::findOrFail($id);
-            $validatedData = $request->only(['title', 'description', 'featured', 'publish_by', 'publish_date']);
+            $validatedData = $request->only(['title', 'banner', 'description', 'featured', 'publish_by', 'publish_date']);
 
             if ($request->filled('publish_date')) {
                 $validatedData['publish_date'] = Carbon::parse($request->publish_date)->format('Y-m-d');
@@ -107,6 +115,9 @@ class BlogDataController extends Controller
             $blog->update($validatedData);
             DB::commit();
             return redirect()->route('blog.list')->with('success', 'Blog updated successfully!');
+        } catch (ValidationException $e) {
+            DB::rollBack();
+            return redirect()->back()->withErrors($e->errors())->withInput();
         } catch (\Exception $e) {
             DB::rollBack();
             return redirect()->back()->with('error', 'Failed to update Blog: ' . $e->getMessage());
@@ -114,12 +125,24 @@ class BlogDataController extends Controller
     }
 
     public function destroy($id) {
-        $blog = Blog::findOrFail($id);
-        if ($blog->banner) {
-            Storage::disk('public')->delete('blog-banner/' . $blog->banner);
+
+        DB::beginTransaction();
+        try {
+            $blog = Blog::findOrFail($id);
+            if ($blog->banner) {
+                Storage::disk('public')->delete('blog-banner/' . $blog->banner);
+            }
+
+            if (!is_null($blog)) {
+                $blog->delete();
+            }
+
+            DB::commit();
+            return redirect()->back()->with('success', 'Blog deleted successfully!');
+        } catch (\Exception $e) {
+            DB::rollBack();
+            return redirect()->back()->with('error', 'Failed to delete Blog: ' . $e->getMessage());
         }
-        $blog->delete();
-        return redirect()->back()->with('success', 'Blog deleted successfully!');
     }
 
     public function view($id) {
