+++ app/Http/Controllers/Portal/testimonialDataController.php
@@ -8,7 +8,7 @@ use Illuminate\Http\Request;
 use Carbon\Carbon;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Facades\DB;
-
+use Illuminate\Validation\ValidationException;
 
 class testimonialDataController extends Controller
 {
@@ -18,25 +18,29 @@ class testimonialDataController extends Controller
     }
 
     public function store(Request $request) {
+
+        $request->validate([
+            'user_id'        => 'nullable|exists:users,id',
+            'name'        => 'required|string|max:255',
+            'heading'        => 'required|string|max:255',
+            'description'      => 'nullable|string',
+            'rating'         => 'nullable|integer|between:1,5',
+            'verified'       => 'nullable|boolean',
+        ]);
+
         DB::beginTransaction();
         try {
-            $request->validate([
-                'user_id'        => 'nullable|exists:users,id',
-                'name'        => 'required|string|max:255',
-                'heading'        => 'required|string|max:255',
-                'description'      => 'nullable|string',
-                'rating'         => 'nullable|integer|between:1,5',
-                'verified'       => 'nullable|boolean',
-            ]);
-
             $validatedData = $request->only(['user_id', 'name', 'heading', 'description', 'rating', 'verified']);
             
             Testimonial::create($validatedData);
             DB::commit();
             return redirect()->route('testimonial.list')->with('success', 'Testimonial submitted successfully!');
+        } catch (ValidationException $e) {
+            DB::rollBack();
+            return redirect()->back()->withErrors($e->errors())->withInput();
         } catch (\Exception $e) {
             DB::rollBack();
-            return redirect()->back()->with('error', 'Failed to submit Blog: ' . $e->getMessage());
+            return redirect()->back()->with('error', 'Failed to update Testimonial: ' . $e->getMessage());
         }
     }
 
@@ -68,16 +72,18 @@ class testimonialDataController extends Controller
     }
 
     public function update(Request $request, $id) {
+
+        $request->validate([
+            'user_id'        => 'nullable|exists:users,id',
+            'name'        => 'required|string|max:255',
+            'heading'        => 'required|string|max:255',
+            'description'      => 'nullable|string',
+            'rating'         => 'nullable|integer|between:1,5',
+            'verified'       => 'nullable|boolean',
+        ]);
+
         DB::beginTransaction();
         try {
-            $request->validate([
-                'user_id'        => 'nullable|exists:users,id',
-                'name'        => 'required|string|max:255',
-                'heading'        => 'required|string|max:255',
-                'description'      => 'nullable|string',
-                'rating'         => 'nullable|integer|between:1,5',
-                'verified'       => 'nullable|boolean',
-            ]);
             
             $testimonial = Testimonial::findOrFail($id);
             $validatedData = $request->only(['user_id', 'name', 'heading', 'description', 'rating', 'verified']);
@@ -85,6 +91,9 @@ class testimonialDataController extends Controller
             $testimonial->update($validatedData);
             DB::commit();
             return redirect()->route('testimonial.list')->with('success', 'Testimonial updated successfully!');
+         } catch (ValidationException $e) {
+            DB::rollBack();
+            return redirect()->back()->withErrors($e->errors())->withInput();
         } catch (\Exception $e) {
             DB::rollBack();
             return redirect()->back()->with('error', 'Failed to update Testimonial: ' . $e->getMessage());
@@ -92,9 +101,21 @@ class testimonialDataController extends Controller
     }
 
     public function destroy($id) {
-        $testimonial = Testimonial::findOrFail($id);
-        $testimonial->delete();
-        return redirect()->back()->with('success', 'Testimonial deleted successfully!');
+
+        DB::beginTransaction();
+        try {
+            $testimonial = Testimonial::findOrFail($id);
+
+            if (!is_null($testimonial)) {
+                $testimonial->delete();
+            }
+
+            DB::commit();
+            return redirect()->back()->with('success', 'Testimonial deleted successfully!.');
+        } catch (\Exception $e) {
+            DB::rollBack();
+            return redirect()->back()->with('error', 'Failed to delete Testimonial: ' . $e->getMessage());
+        }
     }
 
     public function view($id) {
